fix ports for windows
