+++ cmd/submodelrepositoryservice/config/access_rules/example.json
