+++ internal/aasregistry/security_index_tests/aasregistry_security_index_integration_test.go
@@ -51,6 +51,6 @@ func TestIntegration(t *testing.T) {
 func TestMain(m *testing.M) {
 	os.Exit(testenv.RunComposeTestMain(m, testenv.ComposeTestMainOptions{
 		ComposeFile: "docker_compose/docker_compose.yml",
-		HealthURL:   "http://localhost:6004/test/health",
+		HealthURL:   "http://localhost:6005/test/health",
 	}))
 }
