+++ internal/conceptdescriptionrepository/integration_tests/it_config.json
@@ -2,20 +2,20 @@
     {
         "context": "000 - Description endpoint",
         "method": "GET",
-        "endpoint": "http://localhost:6004/description",
+        "endpoint": "http://localhost:6005/description",
         "expectedStatus": 200
     },
     {
         "context": "001 - Get All Concept Descriptions (Empty)",
         "method": "GET",
-        "endpoint": "http://localhost:6004/concept-descriptions",
+        "endpoint": "http://localhost:6005/concept-descriptions",
         "expectedStatus": 200,
         "shouldMatch": "expected/GetAllConceptDescriptionsEmpty.json"
     },
     {
         "context": "002 - Create Concept Description A",
         "method": "POST",
-        "endpoint": "http://localhost:6004/concept-descriptions",
+        "endpoint": "http://localhost:6005/concept-descriptions",
         "expectedStatus": 201,
         "data": "bodies/ConceptDescriptionA.json",
         "shouldMatch": "bodies/ConceptDescriptionA.json"
@@ -23,14 +23,14 @@
     {
         "context": "003 - Get Concept Description A by ID",
         "method": "GET",
-        "endpoint": "http://localhost:6004/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6YTox",
+        "endpoint": "http://localhost:6005/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6YTox",
         "expectedStatus": 200,
         "shouldMatch": "bodies/ConceptDescriptionA.json"
     },
     {
         "context": "004 - Create Concept Description B",
         "method": "POST",
-        "endpoint": "http://localhost:6004/concept-descriptions",
+        "endpoint": "http://localhost:6005/concept-descriptions",
         "expectedStatus": 201,
         "data": "bodies/ConceptDescriptionB.json",
         "shouldMatch": "bodies/ConceptDescriptionB.json"
@@ -38,21 +38,21 @@
     {
         "context": "005 - Get All Concept Descriptions (Limit 1 - Page 1)",
         "method": "GET",
-        "endpoint": "http://localhost:6004/concept-descriptions?limit=1",
+        "endpoint": "http://localhost:6005/concept-descriptions?limit=1",
         "expectedStatus": 200,
         "shouldMatch": "expected/GetAllConceptDescriptionsLimit1Page1.json"
     },
     {
         "context": "006 - Get All Concept Descriptions (Limit 1 - Page 2 via Cursor)",
         "method": "GET",
-        "endpoint": "http://localhost:6004/concept-descriptions?limit=1&cursor=dXJuOmV4YW1wbGU6Y2Q6Yjoy",
+        "endpoint": "http://localhost:6005/concept-descriptions?limit=1&cursor=dXJuOmV4YW1wbGU6Y2Q6Yjoy",
         "expectedStatus": 200,
         "shouldMatch": "expected/GetAllConceptDescriptionsLimit1Page2.json"
     },
     {
         "context": "007 - Update Concept Description B via PUT",
         "method": "PUT",
-        "endpoint": "http://localhost:6004/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6Yjoy",
+        "endpoint": "http://localhost:6005/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6Yjoy",
         "expectedStatus": 200,
         "data": "bodies/ConceptDescriptionBUpdated.json",
         "shouldMatch": "bodies/ConceptDescriptionBUpdated.json"
@@ -60,46 +60,46 @@
     {
         "context": "008 - Get Updated Concept Description B by ID",
         "method": "GET",
-        "endpoint": "http://localhost:6004/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6Yjoy",
+        "endpoint": "http://localhost:6005/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6Yjoy",
         "expectedStatus": 200,
         "shouldMatch": "bodies/ConceptDescriptionBUpdated.json"
     },
     {
         "context": "009 - Delete Concept Description A",
         "method": "DELETE",
-        "endpoint": "http://localhost:6004/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6YTox",
+        "endpoint": "http://localhost:6005/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6YTox",
         "expectedStatus": 204
     },
     {
         "context": "010 - Get Concept Description A by ID (Not Found)",
         "method": "GET",
-        "endpoint": "http://localhost:6004/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6YTox",
+        "endpoint": "http://localhost:6005/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6YTox",
         "expectedStatus": 404
     },
     {
         "context": "011 - Get All Concept Descriptions After Delete A",
         "method": "GET",
-        "endpoint": "http://localhost:6004/concept-descriptions",
+        "endpoint": "http://localhost:6005/concept-descriptions",
         "expectedStatus": 200,
         "shouldMatch": "expected/GetAllConceptDescriptionsAfterDeleteA.json"
     },
     {
         "context": "012 - Delete Concept Description B",
         "method": "DELETE",
-        "endpoint": "http://localhost:6004/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6Yjoy",
+        "endpoint": "http://localhost:6005/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6Yjoy",
         "expectedStatus": 204
     },
     {
         "context": "013 - Get All Concept Descriptions (Empty Again)",
         "method": "GET",
-        "endpoint": "http://localhost:6004/concept-descriptions",
+        "endpoint": "http://localhost:6005/concept-descriptions",
         "expectedStatus": 200,
         "shouldMatch": "expected/GetAllConceptDescriptionsEmpty.json"
     },
     {
         "context": "014 - Create Concept Description Full",
         "method": "POST",
-        "endpoint": "http://localhost:6004/concept-descriptions",
+        "endpoint": "http://localhost:6005/concept-descriptions",
         "expectedStatus": 201,
         "data": "bodies/ConceptDescriptionFull.json",
         "shouldMatch": "bodies/ConceptDescriptionFull.json"
@@ -107,28 +107,28 @@
     {
         "context": "015 - Get Concept Descriptions by idShort",
         "method": "GET",
-        "endpoint": "http://localhost:6004/concept-descriptions?idShort=Temperature",
+        "endpoint": "http://localhost:6005/concept-descriptions?idShort=Temperature",
         "expectedStatus": 200,
         "shouldMatch": "expected/GetAllConceptDescriptionsFilteredFull.json"
     },
     {
         "context": "016 - Get Concept Descriptions by isCaseOf",
         "method": "GET",
-        "endpoint": "http://localhost:6004/concept-descriptions?isCaseOf=0173-1%2302-AAO677%23002",
+        "endpoint": "http://localhost:6005/concept-descriptions?isCaseOf=0173-1%2302-AAO677%23002",
         "expectedStatus": 200,
         "shouldMatch": "expected/GetAllConceptDescriptionsFilteredFull.json"
     },
     {
         "context": "017 - Get Concept Descriptions by dataSpecificationRef",
         "method": "GET",
-        "endpoint": "http://localhost:6004/concept-descriptions?dataSpecificationRef=https%3A%2F%2Fadmin-shell.io%2Faas%2F3%2F0%2FDataSpecificationIEC61360",
+        "endpoint": "http://localhost:6005/concept-descriptions?dataSpecificationRef=https%3A%2F%2Fadmin-shell.io%2Faas%2F3%2F0%2FDataSpecificationIEC61360",
         "expectedStatus": 200,
         "shouldMatch": "expected/GetAllConceptDescriptionsFilteredFull.json"
     },
     {
         "context": "018 - Delete Concept Description Full",
         "method": "DELETE",
-        "endpoint": "http://localhost:6004/concept-descriptions/dXJuOmV4YW1wbGU6Y29uY2VwdC1kZXNjcmlwdGlvbjp0ZW1wZXJhdHVyZTox",
+        "endpoint": "http://localhost:6005/concept-descriptions/dXJuOmV4YW1wbGU6Y29uY2VwdC1kZXNjcmlwdGlvbjp0ZW1wZXJhdHVyZTox",
         "expectedStatus": 204
     }
 ]
