+++ internal/conceptdescriptionrepository/security_tests/conceptdescriptionrepository_security_integration_test.go
@@ -51,7 +51,7 @@ func TestMain(m *testing.M) {
 	os.Exit(testenv.RunComposeTestMain(m, testenv.ComposeTestMainOptions{
 		ComposeFile:     "docker_compose/docker_compose.yml",
 		PreDownBeforeUp: true,
-		HealthURL:       "http://localhost:6004/health",
+		HealthURL:       "http://localhost:6005/health",
 		HealthTimeout:   150 * time.Second,
 	}))
 }
