+++ internal/conceptdescriptionrepository/security_tests/it_config.json
@@ -3,21 +3,21 @@
     "context": "Anonymous description allowed",
     "token": null,
     "method": "GET",
-    "endpoint": "http://localhost:6004/description",
+    "endpoint": "http://localhost:6005/description",
     "expectedStatus": 200
   },
   {
     "context": "Anonymous list denied",
     "token": null,
     "method": "GET",
-    "endpoint": "http://localhost:6004/concept-descriptions",
+    "endpoint": "http://localhost:6005/concept-descriptions",
     "expectedStatus": 403
   },
   {
     "context": "POST viewer-visible concept description - admin",
     "token": { "user": "admin", "password": "pwd" },
     "method": "POST",
-    "endpoint": "http://localhost:6004/concept-descriptions",
+    "endpoint": "http://localhost:6005/concept-descriptions",
     "data": "postBody/cdViewerAllowed.json",
     "shouldMatch": "postBody/cdViewerAllowed.json",
     "expectedStatus": 201
@@ -26,7 +26,7 @@
     "context": "POST viewer-hidden concept description - admin",
     "token": { "user": "admin", "password": "pwd" },
     "method": "POST",
-    "endpoint": "http://localhost:6004/concept-descriptions",
+    "endpoint": "http://localhost:6005/concept-descriptions",
     "data": "postBody/cdViewerHidden.json",
     "shouldMatch": "postBody/cdViewerHidden.json",
     "expectedStatus": 201
@@ -35,7 +35,7 @@
     "context": "GET concept descriptions - viewer filtered",
     "token": { "user": "usera", "password": "pwd" },
     "method": "GET",
-    "endpoint": "http://localhost:6004/concept-descriptions",
+    "endpoint": "http://localhost:6005/concept-descriptions",
     "shouldMatch": "expected/expectedViewerList.json",
     "expectedStatus": 200
   },
@@ -43,7 +43,7 @@
     "context": "GET viewer-visible concept description - viewer",
     "token": { "user": "usera", "password": "pwd" },
     "method": "GET",
-    "endpoint": "http://localhost:6004/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6dmlld2VyOmFsbG93ZWQ",
+    "endpoint": "http://localhost:6005/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6dmlld2VyOmFsbG93ZWQ",
     "shouldMatch": "postBody/cdViewerAllowed.json",
     "expectedStatus": 200
   },
@@ -51,14 +51,14 @@
     "context": "GET viewer-hidden concept description - viewer receives hidden 404",
     "token": { "user": "usera", "password": "pwd" },
     "method": "GET",
-    "endpoint": "http://localhost:6004/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6dmlld2VyOmhpZGRlbg",
+    "endpoint": "http://localhost:6005/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6dmlld2VyOmhpZGRlbg",
     "expectedStatus": 404
   },
   {
     "context": "POST editor-allowed concept description - editor",
     "token": { "user": "userx", "password": "pwd" },
     "method": "POST",
-    "endpoint": "http://localhost:6004/concept-descriptions",
+    "endpoint": "http://localhost:6005/concept-descriptions",
     "data": "postBody/cdEditorAllowed.json",
     "shouldMatch": "postBody/cdEditorAllowed.json",
     "expectedStatus": 201
@@ -67,7 +67,7 @@
     "context": "POST editor-denied concept description - editor",
     "token": { "user": "userx", "password": "pwd" },
     "method": "POST",
-    "endpoint": "http://localhost:6004/concept-descriptions",
+    "endpoint": "http://localhost:6005/concept-descriptions",
     "data": "postBody/cdEditorDenied.json",
     "expectedStatus": 403
   },
@@ -75,7 +75,7 @@
     "context": "POST hidden-existing concept description for old-state checks - admin",
     "token": { "user": "admin", "password": "pwd" },
     "method": "POST",
-    "endpoint": "http://localhost:6004/concept-descriptions",
+    "endpoint": "http://localhost:6005/concept-descriptions",
     "data": "postBody/cdEditorHiddenExisting.json",
     "shouldMatch": "postBody/cdEditorHiddenExisting.json",
     "expectedStatus": 201
@@ -84,7 +84,7 @@
     "context": "PUT invisible existing concept description - editor denied by old-state check",
     "token": { "user": "userx", "password": "pwd" },
     "method": "PUT",
-    "endpoint": "http://localhost:6004/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6ZWRpdG9yOmhpZGRlbi1leGlzdGluZw",
+    "endpoint": "http://localhost:6005/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6ZWRpdG9yOmhpZGRlbi1leGlzdGluZw",
     "data": "postBody/cdEditorHiddenExistingUpdatedAllowed.json",
     "expectedStatus": 403
   },
@@ -92,7 +92,7 @@
     "context": "PUT visible concept description to disallowed state - editor denied by new-state check",
     "token": { "user": "userx", "password": "pwd" },
     "method": "PUT",
-    "endpoint": "http://localhost:6004/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6ZWRpdG9yOnBvc3QtYWxsb3dlZA",
+    "endpoint": "http://localhost:6005/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6ZWRpdG9yOnBvc3QtYWxsb3dlZA",
     "data": "postBody/cdEditorAllowedUpdatedDenied.json",
     "expectedStatus": 403
   },
@@ -100,42 +100,42 @@
     "context": "DELETE invisible concept description - editor denied",
     "token": { "user": "userx", "password": "pwd" },
     "method": "DELETE",
-    "endpoint": "http://localhost:6004/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6ZWRpdG9yOmhpZGRlbi1leGlzdGluZw",
+    "endpoint": "http://localhost:6005/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6ZWRpdG9yOmhpZGRlbi1leGlzdGluZw",
     "expectedStatus": 403
   },
   {
     "context": "DELETE viewer-visible concept description - admin",
     "token": { "user": "admin", "password": "pwd" },
     "method": "DELETE",
-    "endpoint": "http://localhost:6004/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6dmlld2VyOmFsbG93ZWQ",
+    "endpoint": "http://localhost:6005/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6dmlld2VyOmFsbG93ZWQ",
     "expectedStatus": 204
   },
   {
     "context": "DELETE viewer-hidden concept description - admin",
     "token": { "user": "admin", "password": "pwd" },
     "method": "DELETE",
-    "endpoint": "http://localhost:6004/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6dmlld2VyOmhpZGRlbg",
+    "endpoint": "http://localhost:6005/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6dmlld2VyOmhpZGRlbg",
     "expectedStatus": 204
   },
   {
     "context": "DELETE editor-allowed concept description - admin",
     "token": { "user": "admin", "password": "pwd" },
     "method": "DELETE",
-    "endpoint": "http://localhost:6004/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6ZWRpdG9yOnBvc3QtYWxsb3dlZA",
+    "endpoint": "http://localhost:6005/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6ZWRpdG9yOnBvc3QtYWxsb3dlZA",
     "expectedStatus": 204
   },
   {
     "context": "DELETE hidden-existing concept description - admin",
     "token": { "user": "admin", "password": "pwd" },
     "method": "DELETE",
-    "endpoint": "http://localhost:6004/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6ZWRpdG9yOmhpZGRlbi1leGlzdGluZw",
+    "endpoint": "http://localhost:6005/concept-descriptions/dXJuOmV4YW1wbGU6Y2Q6ZWRpdG9yOmhpZGRlbi1leGlzdGluZw",
     "expectedStatus": 204
   },
   {
     "context": "GET concept descriptions - admin cleanup check",
     "token": { "user": "admin", "password": "pwd" },
     "method": "GET",
-    "endpoint": "http://localhost:6004/concept-descriptions",
+    "endpoint": "http://localhost:6005/concept-descriptions",
     "shouldMatch": "expected/expectedEmptyList.json",
     "expectedStatus": 200
   }
